requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
